BUR-70: translations only depend on language of course
