+++ frontend/src/translations/en.json
@@ -131,6 +131,7 @@
         "exercise-session": "Exercise sessions",
         "exercise-session-description": "Lorem ipsum dolor sit amet",
         "exams": "Exams",
-        "exams-description": "Lorem ipsum dolor sit amet"
+        "exams-description": "Lorem ipsum dolor sit amet",
+        "error-fetching": "An error occurred while fetching the course data: "
     }
 }
